+++ modules/src/access/events.rs
@@ -1,6 +1,6 @@
-use odra::{types::Address, Event};
-
 use super::access_control::Role;
+use casper_event_standard::Event;
+use odra::Address;
 
 #[derive(Event, PartialEq, Eq, Debug)]
 pub struct OwnershipTransferred {