+++ modules/src/lib.rs
@@ -1,13 +1,13 @@
 #![cfg_attr(not(test), no_std)]
 #![cfg_attr(not(test), no_main)]
 
-// pub mod access;
-// pub mod erc1155;
+pub mod access;
+pub mod erc1155;
 // pub mod erc1155_receiver;
 // pub mod erc1155_token;
 pub mod erc20;
 // pub mod erc721;
 // pub mod erc721_receiver;
 // pub mod erc721_token;
-// pub mod security;
+pub mod security;
 pub mod wrapped_native;