+++ modules/src/security/errors.rs
@@ -1,8 +1,12 @@
-use odra::execution_error;
+use odra::OdraError;
 
-execution_error! {
-    pub enum Error {
-        PausedRequired => 21_000,
-        UnpausedRequired => 21_001,
+pub enum Error {
+    PausedRequired = 21_000,
+    UnpausedRequired = 21_001
+}
+
+impl From<Error> for OdraError {
+    fn from(error: Error) -> Self {
+        OdraError::user(error as u16)
     }
 }